+++ backend/bang-ggood/src/main/java/com/bang_ggood/question/service/ChecklistQuestionService.java
@@ -2,10 +2,19 @@ package com.bang_ggood.question.service;
 
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
+import com.bang_ggood.question.domain.Category;
 import com.bang_ggood.question.domain.ChecklistQuestion;
+import com.bang_ggood.question.domain.CustomChecklistQuestion;
+import com.bang_ggood.question.domain.Question;
+import com.bang_ggood.question.dto.response.CategoryCustomChecklistQuestionResponse;
+import com.bang_ggood.question.dto.response.CategoryCustomChecklistQuestionsResponse;
+import com.bang_ggood.question.dto.response.CustomChecklistQuestionResponse;
 import com.bang_ggood.question.repository.ChecklistQuestionRepository;
+import com.bang_ggood.question.repository.CustomChecklistQuestionRepository;
+import com.bang_ggood.user.domain.User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -14,9 +23,11 @@ import java.util.Set;
 public class ChecklistQuestionService {
 
     private final ChecklistQuestionRepository checklistQuestionRepository;
+    private final CustomChecklistQuestionRepository customChecklistQuestionRepository;
 
-    public ChecklistQuestionService(ChecklistQuestionRepository checklistQuestionRepository) {
+    public ChecklistQuestionService(ChecklistQuestionRepository checklistQuestionRepository, CustomChecklistQuestionRepository customChecklistQuestionRepository) {
         this.checklistQuestionRepository = checklistQuestionRepository;
+        this.customChecklistQuestionRepository = customChecklistQuestionRepository;
     }
 
     @Transactional
@@ -32,5 +43,30 @@ public class ChecklistQuestionService {
                 throw new BangggoodException(ExceptionCode.QUESTION_DUPLICATED);
             }
         });
+    }
+
+    @Transactional(readOnly = true)
+    public CategoryCustomChecklistQuestionsResponse readAllCustomChecklistQuestions(User user) {
+        List<CustomChecklistQuestion> customChecklistQuestions = customChecklistQuestionRepository.findAllByUser(user);
+        List<CategoryCustomChecklistQuestionResponse> allCategoryCustomChecklistQuestions = getAllCategoryCustomChecklistQuestions(
+                customChecklistQuestions);
+
+        return new CategoryCustomChecklistQuestionsResponse(allCategoryCustomChecklistQuestions);
+    }
+
+    private List<CategoryCustomChecklistQuestionResponse> getAllCategoryCustomChecklistQuestions(
+            List<CustomChecklistQuestion> customChecklistQuestions) {
+        List<CategoryCustomChecklistQuestionResponse> response = new ArrayList<>();
+
+        for (Category category : Category.values()) {
+            List<Question> categoryQuestions = Question.findQuestionsByCategory(category);
+            List<CustomChecklistQuestionResponse> questions = categoryQuestions.stream()
+                    .map(question -> new CustomChecklistQuestionResponse(question,
+                            question.isSelected(customChecklistQuestions)))
+                    .toList();
+            response.add(new CategoryCustomChecklistQuestionResponse(category.getId(), category.getName(), questions));
+        }
+
+        return response;
     }
 }
